Initialize this project!
This is the first of a series of repositories that I created to simplify
how I create demo environments, a.k.a "demo-land".

I wrote this to reduce my dependence on the `tap-test-bed` Jenkins
pipelien that I've been using that tends to fail and expires quicker
than I'd like.

Most of this code was adapted from my work on installing
[TMC-SA on AWS](https://github.com/carlosonunez-vmw/tmc-on-aws).

Signed-off-by: Carlos Nunez <ncarlos@vmware.com>