demo: Easy compliance with RAISE
The Naval Rapid Assess and Incorporate Software Engineering framework is
the framework for vetting software supply chains within the United
States Navy. Having software that is "RAISE-compliant" is becoming a
prerequisite for achieving an Authrity to Operate within the USN, which
is required to run consumer/mission-facing applications within the DOD.

This demo illustrates how TAP makes that possible on every commit to
your code base with Supply Chains.

Highlights include:

- Using `SourceScanner` resources to integrate your software builds to
  your desired SAST/DAST/image scanning tools, like Aquasec and
  SonarQube,
- Leveraging the Tanzu Build Service to create hardened, best-practices
  container images with zero work,
- and more!

Signed-off-by: Carlos Nunez <ncarlos@vmware.com>