[dast] Make `config-writer` friendlier
Out of the box, `config-writer` is a `ClusterTemplate` that seems to be
used as a final task prior to "delivery."

This does not work for Supply Chains that also want to perform DAST.

If DAST is performed prior to this stage, then it will fail (because
nothing has been deployed) and cause the deployment to halt.

If DAST happens after `config-writer` without modifications, it will not
show up in the TAP GUI (since, in 1.6.1, the TAP GUI assumes that the
final stage is a stage that does not produce any outputs, which
`ClusterTemplates` do not).

This commit modifies `config-writer` to be a `ClusterConfigTemplate` to
work around this. These templates expose a "config" of some sort, which
allows DAST to be the final stage. Consequently, the DAST
`ClusterImageTemplate` has been changed to a `ClusterTemplate` so that
it renders as a "final" resource in the TAP GUI.

Signed-off-by: Carlos Nunez <ncarlos@vmware.com>